Lesson 11 completed
Platforms are not moving back and forth between target location and start. the target location is a vector in space that uses a meta specifier so I can move it around. When the  target location is reached values are swap so it goes back.  (We check if the has reached the target by checking the size of the Fvector)